% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_stat.R
\name{batch_stat}
\alias{batch_stat}
\title{Extract and summarize batch impact statistics}
\usage{
batch_stat(inlist, Yv = "conv_area", Cv, BDv, COSTv, Yst, Cst, BDst, COSTst)
}
\arguments{
\item{inlist}{List of impact tables from batch run}

\item{Yv}{agricultural impact metric (new land required)}

\item{Cv}{Carbon impact metric}

\item{BDv}{Biodiversity impact metric}

\item{Yst}{Function to calculate impacts across crops (e.g. sum)}

\item{Cst}{Function to calculate impacts across carbon (e.g. sum)}

\item{BDst}{Function to calculate impacts across biodiversity (e.g. mean)}

\item{COSTst}{Function to calculate cost impact (e.g. mean)}

\item{COSTc}{Cost impact metric}
}
\value{
data.table
}
\description{
Extract and summarize batch impact statistics
}
\keyword{internal}

