% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_length}
\alias{check_length}
\title{Checks and object length to make sure it is equal to 1 or max number of input layers}
\usage{
check_length(x, lx, name = "")
}
\arguments{
\item{x}{object to check}

\item{lx}{length it should be}

\item{name}{Optional name to add in for output message}
}
\description{
Checks and object length to make sure it is equal to 1 or max number of input layers
}
\details{
Internal function that checks whether object is either of length 1 or equal to some longer length,
such as the number of layers in a RasterBrick. If it is not equal to those numbers, it causes a failure.
}
\keyword{internal}

