% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{conv_plot}
\alias{conv_plot}
\title{Plots multiple maps showing cropland conversions}
\usage{
conv_plot(scen, ldim = c(0.4, 0.75), sdim = c(0.03, 0.12, 0.25),
  tdim = c(0.58, 0.18), legdim = c(0.8, 0.15), brks = c(0, 0.01, 0.05,
  0.1, 0.15, 0.2), brks2 = c(0, seq(0.001, 0.601, 0.15)),
  yor = c("transparent", brewer.pal(9, name = "YlOrRd")[5:8]),
  gr = c("transparent", brewer.pal(9, name = "Greens")[5:8]),
  bp = c("grey90", brewer.pal(9, name = "Blues")[c(4, 6, 7, 9)]), cx = 1.2,
  ladj = 1)
}
\arguments{
\item{scen}{List index specification which of several results to read in}
}
\value{
map
}
\description{
Plots multiple maps showing cropland conversions
}
\keyword{internal}

