% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dt_list_to_raster}
\alias{dt_list_to_raster}
\title{Converts rasters to data.tables}
\usage{
dt_list_to_raster(base, inlist, CRSobj)
}
\arguments{
\item{inlist}{List of input data.tables in same format as output by
\code{\link{raster_list_to_dt}}}

\item{CRSobj}{projected coordinate system}
}
\description{
Converts rasters to data.tables
}
\details{
This assumes that rasters are all of equal size, because it creates
a coordinate DT
}
\keyword{internal}

