% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{fname}
\alias{fname}
\title{Filenaming function for rasters}
\usage{
fname(path, code, ext = ".tif")
}
\arguments{
\item{path}{folder and name root for file}

\item{code}{Result of run_code}
}
\description{
Filenaming function for rasters
}
\keyword{internal}

