% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rast_math.R
\name{gdal_calc}
\alias{gdal_calc}
\title{Run gdal_calc.py}
\usage{
gdal_calc(cstr, x, filename, type = "UInt16", gdformat = "GTiff",
  overwrite = TRUE, robject = FALSE)
}
\arguments{
\item{cstr}{String providing the raster calculation to be performed, e.g. "(A == 1) * 10"}

\item{x}{a list of rasters, or filenames of rasters, where list element names match names in cstr}

\item{filename}{output filename to write, with gdal compliant filename extension}

\item{type}{The gdal data type, e.g. "UInt16" (default), "Byte", etc.}

\item{gdformat}{The gdal format for output. Default if "GTiff" for geotiff}

\item{overwrite}{Option to overwrite existing file name. Default is TRUE}

\item{robject}{Name of robject to return. Default is FALSE, meaning raster is simply written to disk}
}
\value{
The calculated raster
}
\description{
Allows gdal_calc.py to be run, with limited options, for arbitrary calculations
}
\note{
This still has to be fixed to handle multi-band rasters gracefully
}
\keyword{internal}

