% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{heat}
\alias{heat}
\title{Create heat map from group of conversion map data.table}
\usage{
heat(bpath, basedt, todisk = TRUE)
}
\arguments{
\item{bpath}{Path to the batch run you want to make}

\item{basedt}{The base (mask) data.table for conversion to raster}

\item{todisk}{TRUE/FALSE, whether to write raster to disk}
}
\value{
A raster, optionally written to disk in the batch folder, containing 
the frequency (in percent) of conversions for each pixel in the map.
}
\description{
Create heat map from group of conversion map data.table
}

