% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{imp_plot}
\alias{imp_plot}
\title{Impact map plot}
\usage{
imp_plot(dtr, impacts, fpath, fnm, ftitle, input_key)
}
\arguments{
\item{dtr}{Conversion raster}

\item{impacts}{data.table of conversion impacts from tradeoff_mod}

\item{fpath}{Directory where you want map image written}

\item{fnm}{Name for the map impact image}

\item{ftitle}{Text string for map title}

\item{input_key}{Location code}
}
\value{
Output map of conversion impacts in png format
}
\description{
Impact map plot
}

