% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{imp_sorter}
\alias{imp_sorter}
\title{Correctly order file names of conversion maps from a batch directory}
\usage{
imp_sorter(bpath, inds = NULL, pattern = "csv", fullnames = FALSE)
}
\arguments{
\item{bpath}{Path to the batch run you want to make}

\item{inds}{Vector of iteration index numbers}

\item{pattern}{csv to get conversion maps (note: argument can be internalized)}
}
\value{
A sorted list of impact table file names
}
\description{
Correctly order file names of conversion maps from a batch directory
}
\note{
This function is useful for analyzing model results and necessary to 
correctly match maps with impact table results.
}

