% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact.R
\name{impact}
\alias{impact}
\title{Calculates impacts of cropland conversions}
\usage{
impact(conv, il)
}
\arguments{
\item{conv}{data.table of cropland conversions}

\item{il}{List of inputs created by input_handler}
}
\description{
Calculates impacts of cropland conversions
}
\details{
Calculates summary statistics of the impacts of cropland 
conversions.
}
\note{
Some estimate of fragmentation might be useful here.
}
\examples{
# To be updated
}

