% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/impact_dt.R
\name{impact_dt}
\alias{impact_dt}
\title{Calculates impacts of cropland conversions}
\usage{
impact_dt(conv, carbon, pot_yield, div_list, cost, crop_frac, cropnames, ha)
}
\arguments{
\item{conv}{data.table of cropland conversions}

\item{carbon}{data.table of veg and soil carbon stocks}

\item{pot_yield}{data.table of potential crop yields (modified as needed)}

\item{div_list}{list of diversity data.tables}

\item{cropnames}{Vector of crop names in analysis}

\item{ha}{Pixel area}
}
\description{
Calculates impacts of cropland conversions
}
\details{
Calculates summary statistics of the impacts of cropland
conversions, including total tons carbon lost, carbon loss/ton crop yield,
total area converted per crop, total area loss of PAs, mean diversity of
converted pixels, mean diversity/ton crop yield
}
\note{
Some estimate of fragmentation might be useful here. I have to fix
impacts to not double-count where winter-crops are grown (wheat here)
}
\examples{
rc <- run_code(input_key = "ZA")
il <- input_handler(input_key = "ZA", ybetas = c(1, 1), code = rc,
                    ybeta_update = 1)
CRSobj <- projection(raster("external/ext_data/ZA-carbon.tif"))
ha <- res(raster("external/ext_data/ZA-carbon.tif"))[2]^2 / 10000
prod_targ <- c("maize" = 4, "cassava" = 2, "ground" = 2, "cotton" = 2,
               "soy" = 2, "pulse" = 2,"sunflower" = 2, "sugarcane" = 2,
               "wheat" = 2)
ybetas <- list(1, 1)
cbetas <- c(1, 1, 0, 0)
names(cbetas) <- c("Ag", "C", "bd", "cost")
impnames <- c("carbon", "richness", "pas")

# target
target <- targets_dt(prod_targ = prod_targ, currprod = il$currprod,
                     potprod = il$pp_curr)
# constraints module
c_prob <- constraints_dt(inlist = il[c("y_std", "carbon_p", "cons_p",
                         "cost")],
                         cbetas = cbetas, code = rc,
                         cropnames = il$cropnames)

# convert module
converted <- convert_dt(conv_prob = c_prob, target = target,
                        crop_frac = il$cropfrac,
                        pot_yield = il$p_yield, cropnames = il$cropnames,
                        base = il$mask, ha = ha, keep_index = FALSE)
# impacts
impacts <- impact_dt(conv = converted,
                     carbon = il$carbon[, c("veg", "soil"), with = FALSE],
                     pot_yield = il$p_yield, div_list = il[impnames[2:3]],
                     crop_frac = il$cropfrac, cropnames = il$cropnames,
                     ha = ha)
}

