% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_modules.R
\name{impact_mod1}
\alias{impact_mod1}
\title{Impacts module 1 (carbon)}
\usage{
impact_mod1(il, conv, conv_area)
}
\arguments{
\item{il}{List of inputs created by input_handler}

\item{conv_area}{matrix of converted area}

\item{conv_mat}{binary matrix of cropland conversions}
}
\value{
data.frame of total C loss and mean carbon loss to yield ratio
}
\description{
Impacts module 1 (carbon)
}
\keyword{internal}

