% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_modules.R
\name{impact_mod2}
\alias{impact_mod2}
\title{Impacts module 2 (cose)}
\usage{
impact_mod2(il, conv, conv_area)
}
\arguments{
\item{il}{List of inputs created by input_handler}

\item{conv_area}{matrix of converted area}

\item{conv_mat}{binary matrix of cropland conversions}
}
\value{
data.table of costs
}
\description{
Impacts module 2 (cose)
}
\keyword{internal}

