% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_modules.R
\name{impact_mod3}
\alias{impact_mod3}
\title{Impacts module 3 (biodiversity)}
\usage{
impact_mod3(il, conv_area)
}
\arguments{
\item{il}{List of inputs created by input_handler}

\item{conv_area}{matrix of converted area}
}
\value{
data.frame of total biodiversity loss, measured as rarity weighted ha
}
\description{
Impacts module 3 (biodiversity)
}
\keyword{internal}

