% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_handler.R
\name{input_handler}
\alias{input_handler}
\title{Data preparation function}
\usage{
input_handler(path, crops, ybetas = list(1, 1), ybeta_update = 0,
  exist_list = NULL, silent = TRUE, code = NULL)
}
\arguments{
\item{path}{Path to data directory}

\item{crops}{Character vector of crop names}

\item{ybetas}{list of either 2 rasters or 2 vectors providing yield 
modifications for climate & irrigation. Defaults to 1, 1, currently inactive}

\item{ybeta_update}{1 or 0; determines whether yield_mod_* is run or not.}

\item{exist_list}{Default is NULL, other an existing list of processed inputs}

\item{silent}{Hide or show print statements describing processing steps 
(TRUE [default] or FALSE)}
}
\value{
List of input variables for model
}
\description{
Data preparation function
}
\details{
This function reads in the necessary raster or data.table inputs for 
tradeoff_mod, but will pass through an existing list if they are provided in 
exist_list. Yield modifications and standardizations are performed when 
ybeta_update = 1 and/or exist_list = NULL. In exist_list is NULL, then the 
full routine will always be run.
}
\note{
It is currently used internally by tradeoff_mod, but could well be 
pulled outside of it.
}
\examples{
# Outdated and removed, need to be added
il <- input_handler(path = "external/data/ZA", crops = "maize")
}

