% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{leg_fun}
\alias{leg_fun}
\title{Legend function for model output plots}
\usage{
leg_fun(ldim, sdim, tdim, legdim, brks, brks2, cx, yor, gr, bp, ladj,
  textcol = "white", bordercol = "white", lp = NULL)
}
\arguments{
\item{ldim}{Long dimensions of legend}

\item{sdim}{Short dimensions of legend}

\item{tdim}{Text x and y}

\item{legdim}{x and y coordinate for farmblock area legend}
}
\value{
Three stacked legends using lmisc::flex_legend, one conventional legend
}
\description{
Legend function for model output plots
}
\keyword{internal}

