% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nm_up}
\alias{nm_up}
\title{Layer assigning name for rasterStackBrick}
\usage{
nm_up(x, namevec)
}
\arguments{
\item{x}{RasterStackBrick}

\item{namevec}{vector of layer names}
}
\description{
Layer assigning name for rasterStackBrick
}
\keyword{internal}

