% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_dominator.R
\name{non_dominator}
\alias{non_dominator}
\title{Remove dominated solutions}
\usage{
non_dominator(intable, cnames, bcode)
}
\arguments{
\item{intable}{Input data.table with impact results}

\item{cnames}{Which parameters were given weight values}
}
\description{
Remove dominated solutions
}
\details{
Provides brute force elimination of dominated solutions
}
\keyword{internal}

