% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{Approximates Pareto front for multi-objective optimization of AgroEcoTradeoff 
Model outputs}
\usage{
pareto(cnames, step = 0.1, prod_targ, yblist = list(1, 1), todisk = TRUE)
}
\arguments{
\item{cnames}{A list of the constraints to optimize over}

\item{step}{The step interval over which to search for optimal solutions}

\item{prod_targ}{Production targets passed as list. See examples for format.}

\item{yblist}{A two element list of yield modifiers (currently disabled)}
}
\description{
This function generates a series of possible outputs from the
AgroEcoTradeoff Model and retains those that are non-dominated, effectively
mimicking the traditional weighted sum method for multi-objective optimization.
}

