% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto_steps}
\alias{pareto_steps}
\title{Generate weight combinations for running pareto function}
\usage{
pareto_steps(cnames, step = 0.1)
}
\arguments{
\item{cnames}{A vector of constraint names to optimize over e.g. c("Y", "BD")}

\item{step}{The step interval over which to search for optimal solutions
#@param yblist A two element list giving yield modifications}
}
\description{
Generate weight combinations for running pareto function
}
\keyword{internal}

