% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rast_math.R
\name{rast_math}
\alias{rast_math}
\title{Evaluates a raster or raster using logical tests}
\usage{
rast_math(x, tmpdir = NULL, format = "GTiff", NAval = NULL,
  filename = NULL)
}
\arguments{
\item{x}{Any raster calculation statement}

\item{tmpdir}{Location raster will be written to if too big to handle in memory}

\item{format}{Default format for rasters, here geotiff}

\item{NAval}{Value to set to NA, default is NULL}

\item{filename}{Output raster name, with gdal compliant extension}
}
\value{
Raster* resulting from raster algebra statement
}
\description{
Evaluates a raster or raster using logical tests
}
\details{
The file writing part is a bit complicated. The name you provide the function will force a write
to disk only if the calculation can be done in memory. Otherwise, if raster wrote a temporary file to
disk because it was too big for memory, the function returns that object.  This is why the function gives
control as to where you can write a filename.
}
\note{
This function was modified from one found in my old ldefuncs.rs.R. As noted there, this should be
improved to run in blocks, and perhaps parallelize, while calculation is being done. In future this
function, if it proves useful, should go into separate R library for raster functions.
}

