% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{raster_list_to_dt}
\alias{raster_list_to_dt}
\title{Converts rasters to data.tables}
\usage{
raster_list_to_dt(inlist, base = TRUE)
}
\arguments{
\item{inlist}{List of input Raster*s}

\item{base}{TRUE (default), in which case data.table with coordinates and
index is created, not if FALSE.}
}
\description{
Converts rasters to data.tables
}
\details{
This assumes that rasters are all of equal size, because it creates
a coordinate DT
}
\keyword{internal}

