% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rast_math.R
\name{rclass_chk}
\alias{rclass_chk}
\title{Checks class of raster object}
\usage{
rclass_chk(x)
}
\arguments{
\item{x}{raster* object to check}
}
\description{
Checks class of raster object
}
\details{
Internal function
}
\keyword{internal}

