% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{run_code}
\alias{run_code}
\title{Creates unique code for outputs from each simulation}
\usage{
run_code(input_key)
}
\arguments{
\item{input_key}{Input key (country code) passed through trademod function}
}
\description{
Creates unique code for outputs from each simulation
}
\note{
This borrow the file-naming bit of raster::rasterTmpFile()
}
\keyword{internal}

