% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_base_path}
\alias{set_base_path}
\title{Gives model base path}
\usage{
set_base_path()
}
\description{
Gives model base path
}
\note{
This sets the working directory for the model to the lowest level
directory containing the model name
}
\keyword{internal}

