% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_length}
\alias{set_length}
\title{Checks and sets object length to correct length}
\usage{
set_length(x, lx = y, name = "")
}
\arguments{
\item{x}{object to check}

\item{lx}{length it should be}

\item{name}{Optional name to add in for output message}
}
\description{
Checks and sets object length to correct length
}
\details{
Internal function that checks whether object is either of length 1 or equal to some longer length,
such as the number of layers in a RasterBrick. If it is between those numbers, it sets the object equal to
the first element.
}
\keyword{internal}

