% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{shhh}
\alias{shhh}
\title{Silences print statements or not}
\usage{
shhh(x, silent)
}
\arguments{
\item{x}{print statement}

\item{silent}{TRUE or FALSE}
}
\description{
Silences print statements or not
}
\keyword{internal}

