% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{standardize}
\alias{standardize}
\title{Function to standardize values from 0-1, for raster or data.table}
\usage{
standardize(x)
}
\arguments{
\item{x}{A data.table}
}
\description{
Function to standardize values from 0-1, for raster or data.table
}
\keyword{internal}

