% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_handler.R
\name{ybeta_rast_to_dt}
\alias{ybeta_rast_to_dt}
\title{Raster ybetas to data.tables}
\usage{
ybeta_rast_to_dt(ybetas, cropnames, base)
}
\arguments{
\item{ybetas}{list of either 2 rasters or 2 vectors providing yield 
modifications for climate & irrigation.}

\item{cropnames}{Vector of cropnames in model}
}
\description{
Raster ybetas to data.tables
}
\keyword{internal}

