% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{fetch_inputs}
\alias{fetch_inputs}
\title{Fetches inputs from file structure for trademod}
\usage{
fetch_inputs(path = "external/ext_data", input_key = "ZA", input = "D")
}
\arguments{
\item{input_key}{A unique file identifier for simulation-specific inputs}
}
\description{
Fetches inputs from file structure for trademod
}
\keyword{internal}

