% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_inputs.R
\name{fetch_inputs}
\alias{fetch_inputs}
\title{Fetches inputs from file structure for trademod}
\usage{
fetch_inputs(path)
}
\arguments{
\item{path}{Base path to data}
}
\description{
Fetches inputs from file structure for trademod
}
\details{
Loads in model inputs
}
\examples{
 
ip <- fetch_inputs(path = "external/data/ZA/")
}

