% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{imp_plot}
\alias{imp_plot}
\title{Impact map plot}
\usage{
imp_plot(dtr, impacts, fpath, fnm, ftitle, crops, stats = TRUE,
  cols = c("orange", "red"), input_key = "ZA", legend = TRUE)
}
\arguments{
\item{dtr}{Conversion raster}

\item{impacts}{data.table of conversion impacts from tradeoff_mod}

\item{fpath}{Directory where you want map image written}

\item{fnm}{Name for the map impact image}

\item{ftitle}{Text string for map title}

\item{crops}{Name of crops whose conversions are being plotted}

\item{stats}{TRUE/FALSE if impact statistics should be plotted on map}

\item{cols}{Vector of map colors to use.}

\item{input_key}{Location code}

\item{legend}{TRUE/FALSE Plot legend added to map or not}
}
\value{
Output map of conversion impacts in png format
}
\description{
Impact map plot
}

