% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact.R
\name{impact}
\alias{impact}
\title{Calculates impacts of cropland conversions}
\usage{
impact(conv, il)
}
\arguments{
\item{conv}{data.table of cropland conversions}

\item{il}{List of inputs created by input_handler}
}
\description{
Calculates impacts of cropland conversions
}
\details{
Calculates summary statistics of the impacts of cropland 
conversions.
}
\note{
Some estimate of fragmentation might be useful here.
}
\examples{
rc <- run_code(input_key = "ZA")  
il <- input_handler(input_key = "ZA", ybetas = c(1, 1), code = rc, 
                    ybeta_update = 1)
CRSobj <- projection(raster("external/ext_data/ZA-carbon.tif"))
prod_targ <- c("maize" = 4,  "soy" = 2)
ybetas <- list(1, 1)
cbetas <- c(1, 1, 0, 0)
names(cbetas) <- c("Y", "C", "BD", "COST")
impnames <- c("carbon", "richness", "pas")

# target
target <- targets(prod_targ = prod_targ, currprod = il$currprod, 
                     potprod = il$pp_curr)
# constraints module 
c_prob <- constraints(inlist = il[c("y_std", "carbon_p", "cons_p", 
                         "cost")], 
                         cbetas = cbetas, code = rc, 
                         cropnames = il$cropnames)

# convert module
converted <- convert(conv_prob = c_prob, target = target, 
                    crop_frac = il$cropfrac, 
                        pot_yield = il$p_yield, cropnames = il$cropnames, 
                        base = il$mask, ha = ha, keep_index = FALSE)
# impacts
impacts <- impact(conv = converted, il = il)
}

