% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{Approximates Pareto front for multi-objective optimization of AgroEcoTradeoff 
Model outputs}
\usage{
pareto(cnames, step = 0.1, yblist, targ)
}
\arguments{
\item{cnames}{A list of the constraints to optimize over}

\item{step}{The step interval over which to search for optimal solutions}

\item{yblist}{A two element list giving yield modifications}

\item{targ}{The target multiplier for production}
}
\description{
This function generates a series of possible outputs from the
AgroEcoTradeoff Model and retains those that are non-dominated, effectively
mimicking the traditional weighted sum method for multi-objective optimization.
}

