% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_base_path}
\alias{set_base_path}
\title{Gives model base path}
\usage{
set_base_path()
}
\description{
Gives model base path
}
\note{
This sets the working directory for the model to the highest level
directory containing the model name. This won't run the model if you set it 
up in a lower level directory of the same name
}
\keyword{internal}

