% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{shhh}
\alias{shhh}
\title{Converts rasters to data.tables}
\usage{
shhh(x, silent)
}
\arguments{
\item{x}{print statement}

\item{silent}{TRUE or FALSE}
}
\description{
Converts rasters to data.tables
}
\keyword{internal}

