% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{standardize}
\alias{standardize}
\title{Function to standardize values from 0-1, for raster or data.table}
\usage{
standardize(x, input = "D")
}
\arguments{
\item{x}{A raster* or data.table}

\item{input}{"D" for data.table (default) or "R" for raster*}
}
\description{
Function to standardize values from 0-1, for raster or data.table
}
\keyword{internal}

