% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_batch.R
\name{tradeoff_batch}
\alias{tradeoff_batch}
\title{Runs tradeoff_mod in batch mode}
\usage{
tradeoff_batch(parms, input_key = "ZA", todisk = FALSE, silent = FALSE,
  ncl = 4)
}
\arguments{
\item{parms}{Matrix or data.frame providing input parameters permutations}

\item{input_key}{Two letter location code indicating input data to load.}

\item{todisk}{TRUE, in which case output data.table is written to disk}

\item{silent}{Silent or verbose model. TRUE (default) or FALSE}
}
\value{
List of tradeoff_mod impact data.frames and conversion data.table if 
todisk is FALSE
}
\description{
Function to drive large numbers of iteration of trademod
}
\note{
This is one place where parallelizing could be useful, replacing for 
loop with foreach.
}
\examples{
targlist <- list(c("maize" = 2, "soy" = 2))
#targlist <- list(c("maize" = 2))
cpmlist <- list(c(1, 1))
cblist <- pareto_steps(c("Y", "C"), step = 0.25)
parms <- batch_params(targlist = targlist, cblist = cblist, 
                      currprodmodlist = list(c(1, 1)))
batch_test <- tradeoff_batch(parms = parms, input_key = "ZA", todisk = FALSE, 
                             silent = TRUE)
}

