% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_mod.R
\name{tradeoff_mod}
\alias{tradeoff_mod}
\title{Spatial tradeoffs model}
\usage{
tradeoff_mod(prod_targ, ybetas, cbetas, currprodmod = 1, input_key = "ZA",
  ybeta_update = 0, exist_list = NULL, silent = TRUE)
}
\arguments{
\item{prod_targ}{Production targets passed as list. See examples for format.}

\item{ybetas}{list of 2 rasters or 2 vectors providing yield modifications}

\item{cbetas}{Vector of constraints to apply to land use. See examples.}

\item{currprodmod}{Modifiers of current yield/production (default 1)}

\item{input_key}{Country/location code indicating input data to load.}

\item{ybeta_update}{1 or 0 (default) - determines whether yield_mod_* is run.}

\item{silent}{Silent or verbose mode (TRUE [default] or FALSE)}
}
\value{
Data.frame of impacts, and data.table referencing conversions.
}
\description{
Master module that calls four major modules and associated 
helper functions, to create a single land use tradeoff scenario.
}
\details{
Much of the detail of running the model, and examples that help to 
illustrate it, are in the examples for the individual model components, 
including checks for consistency between "D" and "R" versions.
}
\note{
The impacts module has to be fixed still to not double-count impacts 
resulting from multi-season cropping in converted pixels.
}
\examples{
prod_targ <- c("maize" = 4, "soy" = 2)

# cbeta vector
cbetas <- c(0.5, 0.5, 0, 0)
names(cbetas) <- c("Y", "C", "BD", "COST")

tdt <- tradeoff_mod("ZA", prod_targ, ybetas, cbetas)  
CRSobj <- projection(raster("external/ext_data/ZA-carbon-priorities.tif"))
plot(dt_to_raster(tdt$conv, CRSobj = CRSobj) - tdr$conv)  # check maps
}

